LPS-64788 Create new class and move code into it
