+++ portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_7_0_1.java
@@ -16,10 +16,10 @@ package com.liferay.portal.upgrade;
 
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.ReleaseInfo;
+import com.liferay.portal.upgrade.v7_0_0.UpgradeNotifications;
 import com.liferay.portal.upgrade.v7_0_1.UpgradeCompany;
 import com.liferay.portal.upgrade.v7_0_1.UpgradeMessageBoards;
 import com.liferay.portal.upgrade.v7_0_1.UpgradeModules;
-import com.liferay.portal.upgrade.v7_0_0.UpgradeNotifications;
 
 /**
  * @author Brian Wing Shun Chan
